+++ source/ComponentHelper.cpp
@@ -26,12 +26,12 @@ void StatComponent::ZeroBonusStats()
 
 int64_t StatComponent::GetHealth() const
 {
-	return ceil(this->m_currentHealth);
+	return (int64_t)ceil(this->m_currentHealth);
 }
 
 int64_t StatComponent::GetMaxHealth() const
 {
-	return ceil(m_baseHealth + m_bonusHealth);
+	return (int64_t)ceil(m_baseHealth + m_bonusHealth);
 }
 
 float StatComponent::GetHealthFraction() const
@@ -41,7 +41,7 @@ float StatComponent::GetHealthFraction() const
 
 float StatComponent::CapHealth()
 {
-	float maxHp = GetMaxHealth();
+	float maxHp = (float)GetMaxHealth();
 	bool contained = m_currentHealth < maxHp;
 
 	// Branchless limit
@@ -112,7 +112,7 @@ float StatComponent::ApplyDamage(const float damage, const bool hitByEnemy)
 
 	// Return current health
 	// Can potentially return negative
-	return GetHealth();
+	return (float)GetHealth();
 }
 
 float StatComponent::ApplyHealing(const float healing, const bool hitByEnemy)
@@ -134,7 +134,7 @@ float StatComponent::ApplyHealing(const float healing, const bool hitByEnemy)
 
 	// Return current health
 	// Can potentially return greater than max, but StatCalcSystem does limit current health to max health when calculating stats later
-	return GetHealth();
+	return (float)GetHealth();
 }
 
 void StatComponent::UpdateBonusHealth(const float delta)
